requirements.txt
Flask==2.0.2
gunicorn==20.1.0
typing-extensions==3.7.4
Werkzeug==2.0.2
tensorflow-cpu==2.9.1

flask-restplus==0.13.0
gunicorn==20.1.0
keras==2.9.0
numpy==1.20
Pillow==8.4.0
requests==2.26.0